fixed math again

+++ src/dsp.h
@@ -21,13 +21,13 @@ inline float tanh_(float x) {
   x = clamp(x, -1.f, 1.f);
   const float twox = x * x;
   x = x + 0.16489087f * twox * x + 0.00985468f * twox * twox * x;
-  return x / sqrtf(1.f + x * x);
+  return x / std::sqrt(1.f + x * x);
 }
 
 // log : [0,1] -> [0,1]
 inline float plog(const float in) {
   constexpr float recip = 1 / 3.f;
-  return recip * log10f(in) + 1.f;
+  return recip * std::log10(in) + 1.f;
 }
 
 } // namespace dsp
